Nodes1D: remove Chebyshev and pure Jacobi Quadrature since functionality is misleading and unused in FEM
